move all bjs utils into module
